backend/requirements.txt
# Nexus Mail Backend - Dependencies
# Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
motor==3.6.0
pymongo==4.9.2

# Redis (task queue)
redis==5.2.1

# Google APIs
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# AI Providers
groq==0.15.0
openai==1.58.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Date/Time
python-dateutil==2.9.0.post0
pytz==2024.2

# HTTP
httpx==0.28.1
aiohttp==3.11.11

# Utilities
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
